requirements.txt
streamlit
numpy
matplotlib
pillow
opencv-python
pandas
plotly
scipy
seaborn
